Add flag for not nullifying the outermost level of member return types
This adds a -Yjava-interop-dont-nullify-outermost flag.
When set, if -Yexplicit-nulls is set as well, we avoid
nullifying the outermost level of a Java field or method.

e.g.

If the flag is not set, we change

Array[String] foo() ==> foo(): Array[String|JavaNull]|JavaNull

If the flag is set, we produce

Array[String|JavaNull]